add ssr support for dashboards
